refactor: BaseIntegrationTest 및 dto 위치 변경
